Time: 4 ms (27.31%), Space: 111.2 MB (95.05%) - LeetHub

+++ 2348-number-of-zero-filled-subarrays/2348-number-of-zero-filled-subarrays.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
-        int cnt = 0;
-        int streak = 0;
+        long long cnt = 0;
+        long long streak = 0;
 
         for(int i=0;i<nums.size();i++){
             if(nums[i]==0){
